debug the crop module
