+++ train/detector/trainner.py
@@ -393,6 +393,7 @@ def singletest(data, net, config, splitfun, combinefun, n_per_run, margin=64, is
         return output
 
 
+
 if __name__ == '__main__':
     main()
     # init_seeds()
